+++ src/app/api/blocks.ts
@@ -6,7 +6,7 @@ import getRandomKey, { SongKey } from "./utils/parameters/getRandomKey";
 import getRandomMode, { SongMode } from "./utils/parameters/getRandomMode";
 import getRandomStartingNote, { SongNote } from "./utils/parameters/getRandomStartingNote";
 import getRandomStyle, { Style } from "./utils/parameters/getRandomStyle";
-import prompt from "./utils/OpenAI/prompt";
+import promptChords from "./utils/OpenAI/promptChords";
 import { uploadS3 } from "./utils/AWS/uploadS3";
 import createMidiTrack from "./utils/Midi/createMidiTrack";
 import parseResponse from "./utils/OpenAI/parse";
@@ -21,13 +21,13 @@ export type Section = {
   signedUrl: string;
 }
 
-async function _generateChords() {
+async function _generateBlocks() {
   const key = getRandomKey();
   const mode = getRandomMode();
   const startingNote = getRandomStartingNote();
   const style = getRandomStyle();
 
-  const response = await prompt(key, mode, startingNote, style);
+  const response = await promptChords(key, mode, startingNote, style);
   if (!response) {
     return null;
   }
@@ -43,7 +43,7 @@ async function _generateChords() {
 }
 
 // Define the server action
-export async function generateChords() {
-  const progression = await _generateChords();
+export async function generateBlocks() {
+  const progression = await _generateBlocks();
   return progression 
 }
