+++ src/app/api/utils/parameters/getRandomStyle.ts
@@ -2,5 +2,6 @@ export type Style = "blues" | "rock" | "jazz" | "hip-hop" | "pop" | "country" | 
 
 export default function getRandomStyle(): Style {
     const styles: Style[] = ["blues", "rock", "jazz", "hip-hop", "pop", "country", "electronic", "midwest emo", "deep house", "berlin techno"];
+    return "deep house"
     return styles[Math.floor(Math.random() * styles.length)];
 }
